+++ pkg/core/tools.go
@@ -1,6 +1,10 @@
 package core
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/gotk3/gotk3/gtk"
+)
 
 // Determine the maximum number of columns.
 func MaxCol() int {
@@ -30,4 +34,68 @@ func MakeRange(min, max int) []int {
 		a[i] = min + i
 	}
 	return a
+}
+
+// If a checkbox is toggled change the filters.
+func SetFilters(checkboxes []*gtk.CheckButton) {
+	for _, cb := range checkboxes {
+		label, _ := cb.GetLabel()
+		switch label {
+		case "EQUAL":
+			if cb.GetActive() {
+				Filters.Equal = true
+			} else {
+				Filters.Equal = false
+			}
+		case "DELETE":
+			if cb.GetActive() {
+				Filters.Delete = true
+			} else {
+				Filters.Delete = false
+			}
+		case "INSERT":
+			if cb.GetActive() {
+				Filters.Insert = true
+			} else {
+				Filters.Insert = false
+			}
+		case "UPDATE":
+			if cb.GetActive() {
+				Filters.Update = true
+			} else {
+				Filters.Update = false
+			}
+		}
+	}
+}
+
+// Initialize checkboxes.
+func InitFilters(i int, cb *gtk.CheckButton) *gtk.CheckButton {
+	switch i {
+	case 0:
+		if Filters.Equal {
+			cb.SetActive(true)
+		} else {
+			cb.SetActive(false)
+		}
+	case 1:
+		if Filters.Delete {
+			cb.SetActive(true)
+		} else {
+			cb.SetActive(false)
+		}
+	case 2:
+		if Filters.Insert {
+			cb.SetActive(true)
+		} else {
+			cb.SetActive(false)
+		}
+	case 3:
+		if Filters.Update {
+			cb.SetActive(true)
+		} else {
+			cb.SetActive(false)
+		}
+	}
+	return cb
 }
